+++ example/abc.py
@@ -1,5 +1,0 @@
-# COMMAND ----------
-a = 2
-print(a)
-
-# COMMAND ----------