+++ src/extension/notebook/models/notebookExecutionContext.ts
@@ -1,0 +1,9 @@
+import { Session } from "../../databricks/models/session";
+
+
+export interface NotebookExecutionContext {
+    notebookId: string
+    language: string
+    session: Session
+    computeId: string
+}