+++ src/extension/ui/sidebar/databricks/models/treeNode.ts
@@ -1,0 +1,11 @@
+import * as vscode from "vscode";
+
+export abstract class TreeNode extends vscode.TreeItem {
+    
+    constructor(
+        label: string,
+        collapsibleState: vscode.TreeItemCollapsibleState = vscode.TreeItemCollapsibleState.None
+    ){
+        super(label, collapsibleState);
+    }
+}