+++ src/extension/ui/sidebar/databricks/nodes/connectionNode.ts
@@ -1,0 +1,17 @@
+import * as vscode from "vscode";
+import { TreeNode } from "../models/treeNode";
+
+export class ConnectionNode extends TreeNode {
+
+    constructor(
+        connected: boolean
+    ) {
+        super(
+            connected ? "Connected" : "Disconnected"
+        );
+
+        this.iconPath = new vscode.ThemeIcon(
+            connected ? "plug" : "debug-disconnect"
+        );
+    }
+}